find_by_id fixes + back image

+++ app/controllers/tags_controller.rb
@@ -5,9 +5,11 @@ respond_to :html, :js
 
 
 def create 
-  task = Task.find(params[:task_id]) 
-  tag = task.tags.create(tag_params)
-  tag.save
+  task = Task.find_by_id(params[:task_id]) 
+  if task != nil
+    tag = task.tags.create(tag_params)
+    tag.save
+  end
   redirect_to :back
 end
 
@@ -26,8 +28,10 @@ end
 
 
 def destroy
-  tag = Tag.find(params[:id]) 
-  tag.destroy
+  tag = Tag.find_by_id(params[:id]) 
+  if tag != nil
+    tag.destroy
+  end
   redirect_to :back
   #respond_to do |format|
   #  format.js { render partial: 'taskslistrefresh'  }
